services/fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Kafka
kafka-python==2.0.2
aiokafka==0.10.0

# MinIO (S3)
minio==7.2.0
boto3==1.29.7

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Image processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# ML
numpy==1.26.2
tensorflow==2.15.0

# Monitoring
prometheus-client==0.19.0

# Utils
python-dotenv==1.0.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
